requirements.txt
# Core ML/AI libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Deep Learning
torch==2.0.1
tensorflow==2.13.0
keras==2.13.1

# Federated Learning
flwr==1.5.0

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
dash>=2.15.0

# MLOps & Experiment Tracking
mlflow==2.6.0
wandb==0.15.9
dvc

# Data Drift Detection
evidently
alibi-detect==0.11.4

# API & Deployment
fastapi==0.101.0
uvicorn==0.23.1
flask==2.3.2
gunicorn==21.2.0

# Containerization
docker==6.1.3

# Data Processing
requests==2.31.0
pyyaml==6.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.19
psycopg2-binary==2.9.7

# Utilities
tqdm==4.65.0
joblib==1.3.1
pyarrow==12.0.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0

